requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
openai>=1.30.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
streamlit>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.0
